fix: compliance checks run again
