+++ package.json
@@ -119,8 +119,8 @@
   "scripts": {
     "clean": "rimraf downloaded/*.yaml generated dist node_modules",
     "test": "aegir lint",
-    "setup:download-spec": "ts-node-esm ./src/download-spec.ts",
-    "build": "run-s setup:download-spec build:aegir",
+    "OLD-setup:download-spec": "ts-node-esm ./src/download-spec.ts",
+    "build": "run-s build:aegir",
     "build:aegir": "aegir build",
     "start": "node dist/src/index.js",
     "release": "npm run build && aegir release --test false --docs false"
