+++ src/checks/get/getAllPins.ts
@@ -1,6 +1,7 @@
 import { ApiCall } from '../../ApiCall.js'
 import type { ServiceAndTokenPair } from '../../types.js'
-import { allPinStatuses, getJoiSchema } from '../../utils/constants.js'
+import { allPinStatuses } from '../../utils/constants.js'
+import { getJoiSchema } from '../../utils/getJoiSchema.js'
 
 const getAllPins = async (pair: ServiceAndTokenPair) => {
   const allPinsApiCall = new ApiCall({
