+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "aegir/src/config/tsconfig.aegir.json",
   "compilerOptions": {
-    "outDir": "dist",
+    "outDir": "dist"
   },
   "include": [
     "test", // remove this line if you don't want to type-check tests
